SerialBot/requirements.txt
pyttsx3==2.90
google-auth==2.23.3
google-cloud-texttospeech==2.14.2
twitchio==2.6.0
python-dotenv==1.0.0
